requirements.txt
pypdf
pinecone
python-dotenv
groq
sentence-transformers
fastapi
uvicorn[standard]